internal/db/pg: check table schemas every app launch
Check tables in 'DB.Prepare' method, not only in tests
