+++ internal/db/pg/pg.go
@@ -111,6 +111,11 @@ func (db *DB) Prepare() error {
 		"old_version": oldVersion, "new_version": newVersion,
 	}).Info("migration process is finished")
 
+	// Check the tables
+	if err := db.checkCreatedTables(); err != nil {
+		return errors.Wrap(err, "database schema is invalid")
+	}
+
 	// Init tables if needed
 	if err = db.initCurrentMonth(); err != nil {
 		const msg = "couldn't init the current month"
@@ -134,6 +139,127 @@ func (db *DB) Prepare() error {
 	// Start cron
 	db.cron.Start()
 
+	return nil
+}
+
+// checkCreatedTables checks tables and their descriptions
+//
+// nolint:funlen
+func (db *DB) checkCreatedTables() error {
+	const tableNumber = 7
+	var n int
+	_, err := db.db.Query(pg.Scan(&n),
+		`SELECT COUNT(DISTINCT table_name) FROM INFORMATION_SCHEMA.COLUMNS WHERE table_schema = 'public'`,
+	)
+	if err != nil {
+		return errors.Wrap(err, "couldn't get number of tables")
+	}
+	if n != tableNumber {
+		return errors.Errorf("invalid number of tables: '%d', expected: '%d'", n, tableNumber)
+	}
+
+	type column struct {
+		Name    string `pg:"column_name"`
+		Type    string `pg:"data_type"`
+		IsNull  bool   `pg:"is_nullable"`
+		Default string `pg:"column_default"`
+	}
+	tables := []struct {
+		name    string
+		columns []column
+	}{
+		// Columns must be sorted by name
+		{
+			name: "months",
+			columns: []column{
+				{Name: "daily_budget", Type: "bigint", Default: "0"},
+				{Name: "id", Type: "bigint", Default: "nextval('months_id_seq'::regclass)"},
+				{Name: "month", Type: "bigint"},
+				{Name: "result", Type: "bigint", Default: "0"},
+				{Name: "total_income", Type: "bigint", Default: "0"},
+				{Name: "total_spend", Type: "bigint", Default: "0"},
+				{Name: "year", Type: "bigint"},
+			},
+		},
+		{
+			name: "days",
+			columns: []column{
+				{Name: "day", Type: "bigint"},
+				{Name: "id", Type: "bigint", Default: "nextval('days_id_seq'::regclass)"},
+				{Name: "month_id", Type: "bigint"},
+				{Name: "saldo", Type: "bigint", Default: "0"},
+			},
+		},
+		{
+			name: "incomes",
+			columns: []column{
+				{Name: "id", Type: "bigint", Default: "nextval('incomes_id_seq'::regclass)"},
+				{Name: "income", Type: "bigint"},
+				{Name: "month_id", Type: "bigint"},
+				{Name: "notes", Type: "text", IsNull: true},
+				{Name: "title", Type: "text"},
+			},
+		},
+		{
+			name: "monthly_payments",
+			columns: []column{
+				{Name: "cost", Type: "bigint"},
+				{Name: "id", Type: "bigint", Default: "nextval('monthly_payments_id_seq'::regclass)"},
+				{Name: "month_id", Type: "bigint"},
+				{Name: "notes", Type: "text", IsNull: true},
+				{Name: "title", Type: "text"},
+				{Name: "type_id", Type: "bigint", IsNull: true},
+			},
+		},
+		{
+			name: "spends",
+			columns: []column{
+				{Name: "cost", Type: "bigint"},
+				{Name: "day_id", Type: "bigint"},
+				{Name: "id", Type: "bigint", Default: "nextval('spends_id_seq'::regclass)"},
+				{Name: "notes", Type: "text", IsNull: true},
+				{Name: "title", Type: "text"},
+				{Name: "type_id", Type: "bigint", IsNull: true},
+			},
+		},
+		{
+			name: "spend_types",
+			columns: []column{
+				{Name: "id", Type: "bigint", Default: "nextval('spend_types_id_seq'::regclass)"},
+				{Name: "name", Type: "text"},
+			},
+		},
+		{
+			name: "migrations",
+			columns: []column{
+				{Name: "created_at", Type: "timestamp with time zone", IsNull: true},
+				{Name: "id", Type: "integer", Default: "nextval('migrations_id_seq'::regclass)"},
+				{Name: "version", Type: "bigint", IsNull: true},
+			},
+		},
+	}
+	var columnsInDB []column
+	for _, table := range tables {
+		_, err := db.db.Query(&columnsInDB,
+			`SELECT column_name, data_type, is_nullable::bool , column_default
+			   FROM INFORMATION_SCHEMA.COLUMNS
+			  WHERE table_name = ?
+			  ORDER BY column_name`, table.name,
+		)
+		if err != nil {
+			return errors.Wrapf(err, "couldn't get description of table '%s'", table.name)
+		}
+
+		err = errors.Wrapf(err, "table has wrong columns: '%+v', expected: '%+v'", columnsInDB, table.columns)
+		if len(table.columns) != len(columnsInDB) {
+			return err
+		}
+		for i := range table.columns {
+			if table.columns[i] != columnsInDB[i] {
+				return err
+			}
+		}
+	}
 	return nil
 }
